Organize imports.  Order Rest-Interface and Queries implemented

+++ orders/orders-service/orders-microservice-ejb/src/main/java/org/cocome/ordersservice/repository/OrderEntryDBRepository.java
@@ -7,6 +7,7 @@ import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.apache.log4j.Logger;
 import org.cocome.ordersservice.domain.OrderEntry;
 
 @Remote
@@ -14,31 +15,67 @@ import org.cocome.ordersservice.domain.OrderEntry;
 public class OrderEntryDBRepository implements OrderEntryRepository {
 	@PersistenceContext(unitName = "InventoryManager")
 	private EntityManager em;
-		
+
+	private static final Logger LOG = Logger.getLogger(OrderEntryDBRepository.class);
+
 	@Override
 	public Long create(OrderEntry entity) {
-		em.persist(entity);
-		return entity.getId();
+		try {
+			em.persist(entity);
+			em.flush();
+			return entity.getId();
+		} catch (Exception e) {
+			LOG.error("DATABASE: Could not create OrderEntry with Id: " + entity.getProductId() + " and OrderId: "
+					+ entity.getOrder().getId());
+			return null;
+		}
 	}
 
 	@Override
 	public OrderEntry find(Long id) {
-		return em.find(OrderEntry.class, id);
+		try {
+			return em.find(OrderEntry.class, id);
+		} catch (Exception e) {
+			LOG.error("DATABASE: Could not find OrderEntry with id: " + id + "     " + e.getMessage());
+			return null;
+		}
+
 	}
 
 	@Override
 	public OrderEntry update(OrderEntry entity) {
-		return em.merge(entity);
+
+		try {
+			return em.merge(entity);
+		} catch (Exception e) {
+			LOG.error("DATABASE: Could not update OrderEntry with Id: " + entity.getId() + " and OrderId: "
+					+ entity.getOrder().getId() + "     " + e.getMessage());
+			return null;
+		}
 	}
 
 	@Override
-	public void delete(Long key) {
-		OrderEntry entity = find(key);
-		em.remove(entity);
+	public boolean delete(Long key) {
+		try {
+			OrderEntry entity = find(key);
+			if (entity != null) {
+				em.remove(entity);
+				return true;
+			}
+
+		} catch (Exception e) {
+			LOG.error("DATABASE: Could not delete OrderEntry with Id: " + key + "     " + e.getMessage());
+		}
+		return false;
 	}
 
 	@Override
 	public Collection<OrderEntry> all() {
-		return em.createQuery("SELECT o FROM OrderEntry o", OrderEntry.class).getResultList();
+		try {
+			return em.createQuery("SELECT o FROM OrderEntry o", OrderEntry.class).getResultList();
+		} catch (Exception e) {
+			LOG.error("DATABASE: Could not retrieve all Order Entries" + "     " + e.getMessage());
+			return null;
+		}
 	}
 }
